Moved and renamed som of the api
