+++ Miracle/include/Miracle/Interface/CurrentApp.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <string_view>
-#include <optional>
 
 #include <Miracle/App.hpp>
 
@@ -35,19 +34,7 @@ namespace Miracle {
 			App::s_currentApp->m_exitCode = exitCode;
 		}
 
-		static inline std::optional<std::u8string_view> getClipboardContent() {
-			if (App::s_currentApp == nullptr) [[unlikely]] throw NoAppRunningError();
-
-			return App::s_currentApp->m_dependencies->getMultimediaFramework().getClipboardContent();
-		}
-
-		static inline void setClipboardContent(const std::u8string_view& content) {
-			if (App::s_currentApp == nullptr) [[unlikely]] throw NoAppRunningError();
-
-			App::s_currentApp->m_dependencies->getMultimediaFramework().setClipboardContent(content);
-		}
-
-		static inline double getTime() {
+		static inline double getRuntime() {
 			if (App::s_currentApp == nullptr) [[unlikely]] throw NoAppRunningError();
 
 			return App::s_currentApp->m_dependencies->getMultimediaFramework().getTime();
